refactor: group flags in flags.go
